Treating midi notes as floats to make them easier to tune, pitchbend, etc.

+++ ggm/ggm.h
@@ -164,7 +164,7 @@ struct midi_rx {
 
 void midi_rx_serial(struct midi_rx *midi, struct usart_drv *serial);
 float midi_scale(uint8_t val, float a, float b);
-float midi_to_frequency(uint8_t note);
+float midi_to_frequency(float note);
 
 //-----------------------------------------------------------------------------
 // events
